+++ src/TeensyDMX.h
@@ -153,6 +153,7 @@ class Receiver final : public TeensyDMX {
           timestamp(0),
           breakPlusMABTime(0),
           breakToBreakTime(0),
+          breakTimestamp(0),
           packetTime(0),
           breakTime(0),
           mabTime(0),
@@ -172,9 +173,10 @@ class Receiver final : public TeensyDMX {
     int size;                   // Packet size
     bool isShort;               // Indicates whether the packet is shorter than
                                 // kMinDMXPacketTime
-    uint32_t timestamp;         // Timestamp, in milliseconds
+    uint32_t timestamp;         // Timestamp at packet completion, in ms
     uint32_t breakPlusMABTime;  // Sum of BREAK and MAB times, in microseconds
     uint32_t breakToBreakTime;  // Time between BREAKs, in microseconds
+    uint32_t breakTimestamp;    // BREAK timestamp, in microseconds
     uint32_t packetTime;        // Packet time, from BREAK start to slot end,
                                 // in microseconds
 
@@ -190,6 +192,7 @@ class Receiver final : public TeensyDMX {
           timestamp(other.timestamp),
           breakPlusMABTime(other.breakPlusMABTime),
           breakToBreakTime(other.breakToBreakTime),
+          breakTimestamp(other.breakTimestamp),
           packetTime(other.packetTime),
           breakTime(other.breakTime),
           mabTime(other.mabTime),
@@ -207,6 +210,7 @@ class Receiver final : public TeensyDMX {
       timestamp = other.timestamp;
       breakPlusMABTime = other.breakPlusMABTime;
       breakToBreakTime = other.breakToBreakTime;
+      breakTimestamp = other.breakTimestamp;
       packetTime = other.packetTime;
       breakTime = other.breakTime;
       mabTime = other.mabTime;
@@ -223,6 +227,7 @@ class Receiver final : public TeensyDMX {
       timestamp = other.timestamp;
       breakPlusMABTime = other.breakPlusMABTime;
       breakToBreakTime = other.breakToBreakTime;
+      breakTimestamp = other.breakTimestamp;
       packetTime = other.packetTime;
       breakTime = other.breakTime;
       mabTime = other.mabTime;